skip rebuild by default, use `-f` flag for force rebuild

+++ x
@@ -3,29 +3,27 @@
 dir="$(dirname "$0")"
 cd "$dir"
 
-if [ -d out ] ; then
-    rm -rf out/*
-else
-    mkdir out
-fi
-
-_main() {
-    arg="${1:-}"
-    mode=draft
-
-    if [ $# -ne 0 ] ; then
-        if [[ "$arg" == d || "$arg" == "draft" ]] ; then
-            mode=draft
-        elif [[ "$arg" == r || "$arg" == "release" ]] ; then
-            mode=release
-        else
-            echo "given wrong argument: \`$arg\`" 1>&2
-            return
-        fi
-    fi
-
-    emacs -Q --script "./build.el" -- "$mode"
+isForceFlag() {
+   [[ "${1:-}" == "-f" || "${1:-}" == "--force" ]]
 }
 
-_main "$@"
+if isForceFlag "${1:-}" || isForceFlag "${2:-}" ; then
+    echo "cleaning up the output directory"
+    if [ -d out ] ; then
+        rm -rf out/*
+    else
+        mkdir out
+    fi
+fi
+
+# FIXME: Passing `"${1:-}" "${2:-}"` causes error after run
+if [ $# -eq 0 ] ; then
+    emacs -Q --script "./build.el"
+elif [ $# -eq 1 ] ; then
+    emacs -Q --script "./build.el" -- "$1"
+elif [ $# -eq 2 ] ; then
+    emacs -Q --script "./build.el" -- "$1" "$2"
+else
+    echo "given too many arguments" 1>&2
+fi
